requirements.txt
fastapi
python-multipart
pillow
mangum
uvicorn